# distillery-server/server/common/api.yml
openapi: 3.0.0
info:
  title: Distillery Controller
  version: '1.0'
  contact:
    name: Jamie Witt
    url: 'https://jamiewitt.com'
    email: me@jamiewitt.com
  description: A server to control distillery functions using Phidget controllers.x
  license:
    name: MIT
servers:
  - url: 'http://localhost:3000/api/v1'
...
paths:
...
  '/distillery/stills/{stillId}':
    get:
      summary: Get a still
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/still'
        '401':
          $ref: '#/components/responses/general-error'
        '403':
          $ref: '#/components/responses/general-error'
        '404':
          $ref: '#/components/responses/general-error'
        '500':
          $ref: '#/components/responses/general-error'
      operationId: 'get-stills-:stillId'
      description: Get a still
      parameters:
        - schema:
            type: string
          in: query
          name: type
          description: 'The type of still to retrieve, it not specified all stills will be returned'
      tags:
        - fractional
        - pot
components:
  schemas:
    simple-message:
      type: object
      title: SimpleMessage
      description: Basic message response
      x-examples:
        example-1:
          message: Operation completed successfully
      additionalProperties: false
      properties:
        message:
          type: string
          example: Operation completed successfully
      required:
        - message
...
    still:
      title: still
      type: object
      description: Describes a still
      properties:
        id:
          type: string
        type:
          type: string
          enum:
            - fractional
            - pot
        status:
          type: string
          enum:
            - free
            - busy
        phidgets:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
                enum:
                  - temperature-probe
                  - solenoid
                  - arm-extend
                  - heating-element
                  - high-voltage-heating-element
              state:
                type: number
      required:
        - id
        - type
...
  securitySchemes:
    x-client-key:
      type: apiKey
      in: header
      description: ''
      name: x-client-key
    token:
      type: openIdConnect
      openIdConnectUrl: 'https://localhost:3000/'
...
  responses:
    general-error:
      description: Example response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/simple-message'
          examples:
            example-1:
              value:
                message: Something went wrong
        application/xml:
          schema:
            type: object
            properties: {}